+++ condition/condition.gd
@@ -48,6 +48,10 @@ const Signals := preload("../event/signal.gd")
 
 # -- INITIALIZATION ------------------------------------------------------------------ #
 
+@warning_ignore("UNUSED_PRIVATE_CLASS_VARIABLE")
+# gdlint:ignore=class-definitions-order
+static var _logger := StdLogger.create(&"std/condition")
+
 var _is_currently_allowed: bool = false
 
 # -- ENGINE METHODS (OVERRIDES) ------------------------------------------------------ #