+++ input/cursor.gd
@@ -31,6 +31,7 @@ static var _logger := StdLogger.create("std/input/cursor")
 var _cursor_captured: bool = false
 var _cursor_confined: bool = false
 var _cursor_visible: bool = false
+var _focus_root: Control = null
 var _hide_actions := PackedStringArray()
 var _hide_actions_if_hovered := PackedStringArray()
 var _hide_delay: float = 0.0
@@ -67,6 +68,30 @@ func hide_cursor() -> void:
 ## there isn't currently a focused node.
 func report_focus_handler_visible(_handler: StdInputCursorFocusHandler) -> void:
 	if not _cursor_visible:
+		_update_focus()
+
+
+## set_focus_root restricts UI focus to be under the scene subtree rooted at `root`. If
+## set, current and future focus handlers will be ignored if they aren't descendants of
+## `root`. Call this with `null` to unset the focus root.
+func set_focus_root(root: Control = null) -> void:
+	assert(
+		root == null or root.is_visible_in_tree(),
+		"invalid argument; root node isn't visible",
+	)
+
+	(
+		_logger
+		. debug(
+			"Updating focus root.",
+			{&"root": root.get_path() if root else ^"/root"},
+		)
+	)
+
+	var changed := _focus_root != root
+	_focus_root = root
+
+	if changed and not _cursor_visible:
 		_update_focus()
 
 
@@ -171,7 +196,6 @@ func update_configuration(action_sets: Array[StdInputActionSet] = []) -> void:
 
 func _exit_tree() -> void:
 	StdGroup.with_id(GROUP_INPUT_CURSOR).remove_member(self)
-	Signals.disconnect_safe(get_viewport().gui_focus_changed, _on_gui_focus_changed)
 
 
 func _input(event: InputEvent) -> void:
@@ -262,7 +286,13 @@ func _update_focus() -> void:
 		get_viewport().gui_release_focus()
 		return
 
-	if get_viewport().gui_get_focus_owner():
+	# When switching scenes, nodes will be rebuilt. Ensure the cached focus root is
+	# still valid before trying to use it.
+	if _focus_root and not is_instance_valid(_focus_root):
+		set_focus_root(null)
+
+	var current_focus := get_viewport().gui_get_focus_owner()
+	if current_focus and (not _focus_root or _focus_root.is_ancestor_of(current_focus)):
 		return
 
 	if _hovered:
@@ -272,12 +302,14 @@ func _update_focus() -> void:
 		_hovered.grab_focus()
 		unset_hovered(_hovered)
 	else:
-		var focus_target := StdInputCursorFocusHandler.get_focus_target()
+		var focus_target := StdInputCursorFocusHandler.get_focus_target(_focus_root)
 		if focus_target:
 			# NOTE: Use a deferred call here so that the current input event gets
 			# swallowed. That ensures the anchor is focused and not a potential
 			# neighbor (depending on what input triggered the change).
 			focus_target.call_deferred(&"grab_focus")
+		elif current_focus:
+			current_focus.release_focus()
 
 
 # -- SIGNAL HANDLERS ----------------------------------------------------------------- #